refactor: operation layout
