+++ packages/api-reference/index.html
@@ -23,7 +23,7 @@
         //   targetKey: 'node',
         //   clientKey: 'undici',
         // },
-        // layout: 'classic'
+        layout: 'classic',
       }
 
       document.getElementById('api-reference').dataset.configuration =
