+++ packages/api-reference/src/components/Content/Operation/OperationResponses.vue
@@ -4,7 +4,9 @@ import type { TransformedOperation } from '@scalar/types/legacy'
 import { useResponses } from '../../../hooks'
 import ParameterList from './ParameterList.vue'
 
-const props = defineProps<{ operation: TransformedOperation }>()
+const props = defineProps<{
+  operation: TransformedOperation
+}>()
 
 const { responses } = useResponses(props.operation)
 </script>
