+++ packages/api-reference/src/components/Content/Operation/index.ts
@@ -1,2 +1,1 @@
-export { default as Operation } from './Operation.vue'
-export { default as OperationAccordion } from './OperationAccordion.vue'
+export { default as Operation } from './components/Operation.vue'
